doc: Revise Adjusting quality.
x264 and x265 recommended quality settings are now the same, consistent with the new official presets. There is still variation between the two encoders and it is generally unhelpful to compare the two in terms of rate factor, but they seem to be fairly close when creating higher quality encodes from film-like sources.